backend/pyproject.toml
[project]
name = "StyleO-backend"
version = "1.0.0"
description = "Backend for StyleO application"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.3",
    "beanie>=2.0.1",
    "taskiq>=0.11.0",
    "taskiq-redis>=0.5.0",
    "taskiq-fastapi>=0.3.0",
    "email-validator>=2.0.0",
    "fastapi>=0.128.0",
    "fastapi-cache2>=0.2.2",
    "google-auth>=2.0.0",
    "google-genai>=1.0.0",
    "passlib[argon2]>=1.7.4",
    "pydantic-settings>=2.0.0",
    "pyjwt>=2.11.0",
    "pymongo>=4.16.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "redis>=7.2.0",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
    "rembg[gpu]>=2.0.73",
    "onnxruntime-gpu>=1.24.3",
    "voyageai>=0.3.7",
]
